Combine TeamTable and TeamList components
