+++ src/components/teams/TeamList/index.js
@@ -1,7 +1,47 @@
 import React, { Component } from 'react'
 import { connect } from 'react-redux'
+import { Table } from 'reactstrap'
 import actions, { teams } from 'modules/team'
-import { TeamTable } from 'components'
+import { TeamLink, UserLink } from 'components'
+
+const TeamTable = ({ teams }) => (
+  <div>
+    <Table>
+      <thead>
+        <tr>
+          <th>Name</th>
+          <th>Lead</th>
+          <th>Members</th>
+        </tr>
+      </thead>
+      <tbody>
+        {teams.map((team, i) =>
+          <tr key={i}>
+            <td>
+              <TeamLink team={team}>{team.name}</TeamLink>
+            </td>
+            <td>
+              {team.lead &&
+                <UserLink user={team.lead}>{team.lead.username}</UserLink>
+              }
+            </td>
+            <td>
+              {team.members &&
+                <ul>
+                  {team.members.map((member, j) =>
+                    <li key={j}>
+                      <UserLink user={member}>{member.username}</UserLink>
+                    </li>
+                  )}
+                </ul>
+              }
+            </td>
+          </tr>
+        )}
+      </tbody>
+    </Table>
+  </div>
+)
 
 class TeamList extends Component {
   componentDidMount() {
